Split SFMT part
